backend/requirements.txt
fastapi
uvicorn
networkx
python-multipart
requests
pillow
python-dotenv
pydantic
httpx
neo4j
simple-icd-10-cm
langgraph>=0.2.0
langchain-core>=0.3.0
streamlit>=1.38.0
pandas>=2.1.0
soundfile>=0.12.0
# Tier 2: CheXNet image analysis (DenseNet-121, NIH ChestX-ray14 weights)
# torch/torchvision installed via Dockerfile using CPU-only wheel index
# torchxrayvision wraps the model; scikit-image is its image-processing dep
torchxrayvision>=1.0.0
scikit-image>=0.21.0
